requirements.txt
Faker==4.1.1
mysql-connector-python==8.0.20
python-dotenv==0.13.0
